Refactor database interactions for unified access
Introduced `App\Core\Database` as a singleton class to manage the PDO database connection centrally. This class uses configuration from `config/config.php`.

Introduced `App\Core\BaseModel` as an abstract base class that all models can extend. Its constructor initializes a protected `$db` property using the `App\Core\Database` singleton, providing a consistent way for models to access the database connection.

Refactored `User`, `Client`, and `OrderModel` to extend `BaseModel` and utilize the inherited `$this->db` for all database operations. Removed their previous individual and inconsistent database connection logic.

This change centralizes database connection management, improving maintainability and consistency of the data access layer.
